refactor block-runtime metadata-core boundaries
