+++ .sampo/changesets/issue-349-metadata-core-boundaries.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/block-runtime: patch
+---
+
+Split `metadata-core` artifact sync routines and endpoint client rendering into focused helper modules without changing the public block-runtime API.
