+++ packages/wp-typia-block-runtime/src/metadata-core.ts
@@ -1,42 +1,21 @@
 import type {} from './typia-tags.js';
 
-import * as fs from 'node:fs';
-import * as path from 'node:path';
-
 import {
-  type GeneratedArtifactFile,
-  type GeneratedArtifactDriftIssue,
   normalizeSyncBlockMetadataFailure,
-  reconcileGeneratedArtifacts,
   resolveSyncBlockMetadataPaths,
 } from './metadata-core-artifacts.js';
 import {
-  assertValidClientIdentifier,
-  normalizeSyncEndpointClientOptions,
-  normalizeSyncRestOpenApiOptions,
-  reserveUniqueClientTypeIdentifier,
-  resolveEndpointClientContract,
-  toJavaScriptStringLiteral,
-  toModuleImportPath,
-  toValidatorAccessExpression,
-} from './metadata-core-endpoint-client.js';
-import { renderPhpValidator } from './metadata-php-render.js';
-import { analyzeSourceType, analyzeSourceTypes } from './metadata-parser.js';
-import {
-  createBlockJsonAttribute,
-  createExampleValue,
-  createManifestDocument,
-  validateWordPressExtractionAttributes,
-} from './metadata-projection.js';
-import {
-  buildEndpointOpenApiDocument,
   type EndpointOpenApiEndpointDefinition,
-  manifestToJsonSchema,
-  manifestToOpenApi,
-  normalizeEndpointAuthDefinition,
   type OpenApiInfo,
-  projectJsonSchemaDocument,
 } from './schema-core.js';
+import {
+  syncEndpointClientModule,
+} from './metadata-core-client-render.js';
+import {
+  syncBlockMetadataArtifacts,
+  syncRestOpenApiArtifacts,
+  syncTypeSchemaArtifacts,
+} from './metadata-core-sync-routines.js';
 
 export interface SyncBlockMetadataOptions {
   blockJsonFile: string;
@@ -338,121 +317,7 @@ export async function syncBlockMetadata(
   options: SyncBlockMetadataOptions,
   executionOptions: ArtifactSyncExecutionOptions = {},
 ): Promise<SyncBlockMetadataResult> {
-  const { blockJsonPath, jsonSchemaPath, manifestPath, openApiPath, phpValidatorPath } =
-    resolveSyncBlockMetadataPaths(options);
-  const { rootNode } = analyzeSourceType(options);
-
-  if (rootNode.kind !== 'object' || rootNode.properties === undefined) {
-    throw new Error(
-      `Source type "${options.sourceTypeName}" must resolve to an object shape`,
-    );
-  }
-  validateWordPressExtractionAttributes(rootNode.properties);
-
-  const driftIssues: GeneratedArtifactDriftIssue[] = [];
-  const lossyProjectionWarnings: string[] = [];
-  let blockJsonArtifact: GeneratedArtifactFile | null = null;
-
-  if (fs.existsSync(blockJsonPath)) {
-    const blockJson = JSON.parse(
-      fs.readFileSync(blockJsonPath, 'utf8'),
-    ) as Record<string, unknown>;
-
-    blockJson.attributes = Object.fromEntries(
-      Object.entries(rootNode.properties).map(([key, node]) => [
-        key,
-        createBlockJsonAttribute(node, lossyProjectionWarnings),
-      ]),
-    );
-    blockJson.example = {
-      attributes: Object.fromEntries(
-        Object.entries(rootNode.properties).map(([key, node]) => [
-          key,
-          createExampleValue(node, key),
-        ]),
-      ),
-    };
-
-    blockJsonArtifact = {
-      content: JSON.stringify(blockJson, null, '\t'),
-      path: blockJsonPath,
-    };
-  } else if (executionOptions.check === true) {
-    driftIssues.push({
-      path: blockJsonPath,
-      reason: 'missing',
-    });
-  } else {
-    fs.readFileSync(blockJsonPath, 'utf8');
-  }
-
-  if (blockJsonArtifact === null) {
-    Object.values(rootNode.properties).forEach((node) => {
-      createBlockJsonAttribute(node, lossyProjectionWarnings);
-    });
-  }
-
-  const manifest = createManifestDocument(
-    options.sourceTypeName,
-    rootNode.properties,
-  );
-  const manifestContent = JSON.stringify(manifest, null, '\t');
-  const jsonSchemaContent = jsonSchemaPath
-    ? JSON.stringify(manifestToJsonSchema(manifest as never), null, '\t')
-    : null;
-  const openApiContent = openApiPath
-    ? JSON.stringify(
-        manifestToOpenApi(manifest as never, {
-          title: options.sourceTypeName,
-        }),
-        null,
-        '\t',
-      )
-    : null;
-  const phpValidator = renderPhpValidator(manifest);
-
-  reconcileGeneratedArtifacts(
-    [
-      ...(blockJsonArtifact ? [blockJsonArtifact] : []),
-      {
-        content: manifestContent,
-        path: manifestPath,
-      },
-      ...(jsonSchemaContent && jsonSchemaPath
-        ? [
-            {
-              content: jsonSchemaContent,
-              path: jsonSchemaPath,
-            },
-          ]
-        : []),
-      ...(openApiContent && openApiPath
-        ? [
-            {
-              content: openApiContent,
-              path: openApiPath,
-            },
-          ]
-        : []),
-      {
-        content: phpValidator.source,
-        path: phpValidatorPath,
-      },
-    ],
-    executionOptions,
-    driftIssues,
-  );
-
-  return {
-    attributeNames: Object.keys(rootNode.properties),
-    blockJsonPath,
-    ...(jsonSchemaPath ? { jsonSchemaPath } : {}),
-    lossyProjectionWarnings: [...new Set(lossyProjectionWarnings)].sort(),
-    manifestPath,
-    ...(openApiPath ? { openApiPath } : {}),
-    phpGenerationWarnings: [...new Set(phpValidator.warnings)].sort(),
-    phpValidatorPath,
-  };
+  return syncBlockMetadataArtifacts(options, executionOptions);
 }
 
 /**
@@ -524,58 +389,7 @@ export async function syncTypeSchemas(
   options: SyncTypeSchemaOptions,
   executionOptions: ArtifactSyncExecutionOptions = {},
 ): Promise<SyncTypeSchemaResult> {
-  const { projectRoot, rootNode } = analyzeSourceType(options);
-  if (rootNode.kind !== 'object' || rootNode.properties === undefined) {
-    throw new Error(
-      `Source type "${options.sourceTypeName}" must resolve to an object shape for schema generation`,
-    );
-  }
-
-  const manifest = createManifestDocument(
-    options.sourceTypeName,
-    rootNode.properties,
-  );
-
-  const jsonSchemaPath = path.resolve(projectRoot, options.jsonSchemaFile);
-  const openApiPath = options.openApiFile
-    ? path.resolve(projectRoot, options.openApiFile)
-    : undefined;
-  reconcileGeneratedArtifacts(
-    [
-      {
-        content: JSON.stringify(
-          projectJsonSchemaDocument(manifestToJsonSchema(manifest as never), {
-            profile: 'rest',
-          }),
-          null,
-          '\t',
-        ),
-        path: jsonSchemaPath,
-      },
-      ...(openApiPath
-        ? [
-            {
-              content: JSON.stringify(
-                manifestToOpenApi(
-                  manifest as never,
-                  options.openApiInfo ?? { title: options.sourceTypeName },
-                ),
-                null,
-                '\t',
-              ),
-              path: openApiPath,
-            },
-          ]
-        : []),
-    ],
-    executionOptions,
-  );
-
-  return {
-    jsonSchemaPath,
-    openApiPath,
-    sourceTypeName: options.sourceTypeName,
-  };
+  return syncTypeSchemaArtifacts(options, executionOptions);
 }
 
 /**
@@ -588,67 +402,7 @@ export async function syncRestOpenApi(
   options: SyncRestOpenApiOptions,
   executionOptions: ArtifactSyncExecutionOptions = {},
 ): Promise<SyncRestOpenApiResult> {
-  const { manifest, openApiPath, projectRoot, typesFile } =
-    normalizeSyncRestOpenApiOptions(options);
-  const sourceTypeNames = Object.values(manifest.contracts).map(
-    (contract) => contract.sourceTypeName,
-  );
-  const analyzedTypes = analyzeSourceTypes(
-    {
-      projectRoot,
-      typesFile,
-    },
-    sourceTypeNames,
-  );
-  const contracts = Object.fromEntries(
-    Object.entries(manifest.contracts).map(([contractKey, contract]) => {
-      const rootNode = analyzedTypes[contract.sourceTypeName];
-      if (rootNode.kind !== 'object' || rootNode.properties === undefined) {
-        throw new Error(
-          `Source type "${contract.sourceTypeName}" must resolve to an object shape for REST OpenAPI generation`,
-        );
-      }
-
-      return [
-        contractKey,
-        {
-          document: createManifestDocument(
-            contract.sourceTypeName,
-            rootNode.properties,
-          ),
-          ...(typeof contract.schemaName === 'string' &&
-          contract.schemaName.length > 0
-            ? { schemaName: contract.schemaName }
-            : {}),
-        },
-      ];
-    }),
-  );
-  reconcileGeneratedArtifacts(
-    [
-      {
-        content: JSON.stringify(
-          buildEndpointOpenApiDocument({
-            contracts,
-            endpoints: manifest.endpoints,
-            info: manifest.info,
-          }),
-          null,
-          '\t',
-        ),
-        path: openApiPath,
-      },
-    ],
-    executionOptions,
-  );
-
-  return {
-    endpointCount: manifest.endpoints.length,
-    openApiPath,
-    schemaNames: Object.values(manifest.contracts).map(
-      (contract) => contract.schemaName ?? contract.sourceTypeName,
-    ),
-  };
+  return syncRestOpenApiArtifacts(options, executionOptions);
 }
 
 /**
@@ -661,322 +415,5 @@ export async function syncEndpointClient(
   options: SyncEndpointClientOptions,
   executionOptions: ArtifactSyncExecutionOptions = {},
 ): Promise<SyncEndpointClientResult> {
-  const { clientPath, manifest, projectRoot, typesFile, validatorsFile } =
-    normalizeSyncEndpointClientOptions(options);
-  analyzeSourceTypes(
-    { projectRoot, typesFile },
-    [...new Set(Object.values(manifest.contracts).map((contract) => contract.sourceTypeName))],
-  );
-  const operationIds = new Set<string>();
-  const importedTypeNames = new Set<string>();
-  const endpointLines: string[] = [];
-  const inlineHelpers = new Set<string>();
-  const validatorPropertyNames = new Map<string, string>();
-  const hasCombinedRequestEndpoints = manifest.endpoints.some(
-    (endpoint) => Boolean(endpoint.bodyContract && endpoint.queryContract),
-  );
-  const occupiedIdentifiers = new Set([
-    'apiValidators',
-    'callEndpoint',
-    'createEndpoint',
-    ...(manifest.endpoints.some(
-      (endpoint) => !endpoint.bodyContract && !endpoint.queryContract,
-    )
-      ? ['validateNoRequest']
-      : []),
-    ...(hasCombinedRequestEndpoints ? ['validateCombinedRequest'] : []),
-  ]);
-
-  for (const endpoint of manifest.endpoints) {
-    const normalizedAuth = normalizeEndpointAuthDefinition(endpoint);
-    const endpointConstantName = `${endpoint.operationId}Endpoint`;
-    assertValidClientIdentifier(endpoint.operationId, 'operationId');
-    assertValidClientIdentifier(endpointConstantName, 'endpoint constant');
-    if (operationIds.has(endpoint.operationId)) {
-      throw new Error(
-        `Duplicate endpoint operationId "${endpoint.operationId}" detected while generating the endpoint client.`,
-      );
-    }
-    for (const identifier of [endpoint.operationId, endpointConstantName]) {
-      if (occupiedIdentifiers.has(identifier)) {
-        throw new Error(
-          `Generated endpoint client identifier "${identifier}" collides with another emitted symbol.`,
-        );
-      }
-    }
-    operationIds.add(endpoint.operationId);
-    occupiedIdentifiers.add(endpoint.operationId);
-    occupiedIdentifiers.add(endpointConstantName);
-
-    const queryContractKey = endpoint.queryContract ?? null;
-    const bodyContractKey = endpoint.bodyContract ?? null;
-    const hasRequest = Boolean(queryContractKey || bodyContractKey);
-    const responseContract = resolveEndpointClientContract(
-      manifest,
-      endpoint.responseContract,
-      endpoint.operationId,
-      'responseContract',
-    );
-    importedTypeNames.add(responseContract.sourceTypeName);
-
-    let requestTypeName = 'undefined';
-    let requestValidatorExpression = 'validateNoRequest';
-    let requestLocationExpression: string | null = null;
-    const queryContract = queryContractKey
-      ? resolveEndpointClientContract(
-          manifest,
-          queryContractKey,
-          endpoint.operationId,
-          'queryContract',
-        )
-      : null;
-    const bodyContract = bodyContractKey
-      ? resolveEndpointClientContract(
-          manifest,
-          bodyContractKey,
-          endpoint.operationId,
-          'bodyContract',
-        )
-      : null;
-
-    if (queryContract && bodyContract) {
-      const queryValidatorExpression = toValidatorAccessExpression(
-        queryContractKey!,
-        validatorPropertyNames,
-      );
-      const bodyValidatorExpression = toValidatorAccessExpression(
-        bodyContractKey!,
-        validatorPropertyNames,
-      );
-      requestTypeName = `{ query: ${queryContract.sourceTypeName}; body: ${bodyContract.sourceTypeName} }`;
-      requestValidatorExpression = `(input) => validateCombinedRequest( input, ${queryValidatorExpression}, ${bodyValidatorExpression} )`;
-      requestLocationExpression = "'query-and-body'";
-      importedTypeNames.add(queryContract.sourceTypeName);
-      importedTypeNames.add(bodyContract.sourceTypeName);
-      inlineHelpers.add('validateCombinedRequest');
-    } else if (queryContract) {
-      requestTypeName = queryContract.sourceTypeName;
-      requestValidatorExpression = toValidatorAccessExpression(
-        queryContractKey!,
-        validatorPropertyNames,
-      );
-      requestLocationExpression = "'query'";
-      importedTypeNames.add(queryContract.sourceTypeName);
-    } else if (bodyContract) {
-      requestTypeName = bodyContract.sourceTypeName;
-      requestValidatorExpression = toValidatorAccessExpression(
-        bodyContractKey!,
-        validatorPropertyNames,
-      );
-      requestLocationExpression = "'body'";
-      importedTypeNames.add(bodyContract.sourceTypeName);
-    } else {
-      inlineHelpers.add('validateNoRequest');
-    }
-
-    const returnCallExpression = hasRequest
-      ? `callEndpoint( ${endpoint.operationId}Endpoint, request, options )`
-      : `callEndpoint( ${endpoint.operationId}Endpoint, undefined, options )`;
-    const returnCallLines =
-      returnCallExpression.length <= 68
-        ? [`\treturn ${returnCallExpression};`]
-        : [
-            `\treturn callEndpoint(`,
-            `\t\t${endpoint.operationId}Endpoint,`,
-            `\t\t${hasRequest ? 'request' : 'undefined'},`,
-            `\t\toptions`,
-            `\t);`,
-          ];
-
-    endpointLines.push(
-      [
-        `export const ${endpointConstantName} = createEndpoint<`,
-        `\t${requestTypeName},`,
-        `\t${responseContract.sourceTypeName}`,
-        `>( {`,
-        `\tauthIntent: ${toJavaScriptStringLiteral(normalizedAuth.auth)},`,
-        ...(normalizedAuth.authMode
-          ? [`\tauthMode: ${toJavaScriptStringLiteral(normalizedAuth.authMode)},`]
-          : []),
-        `\tmethod: ${toJavaScriptStringLiteral(endpoint.method)},`,
-        `\toperationId: ${toJavaScriptStringLiteral(endpoint.operationId)},`,
-        `\tpath: ${toJavaScriptStringLiteral(endpoint.path)},`,
-        ...(requestLocationExpression
-          ? [`\trequestLocation: ${requestLocationExpression},`]
-          : []),
-        `\tvalidateRequest: ${requestValidatorExpression},`,
-        `\tvalidateResponse: ${toValidatorAccessExpression(
-          endpoint.responseContract,
-          validatorPropertyNames,
-        )},`,
-        `} );`,
-        '',
-        `export function ${endpoint.operationId}(`,
-        ...(hasRequest ? [`\trequest: ${requestTypeName},`] : []),
-        `\toptions: EndpointCallOptions`,
-        `) {`,
-        ...returnCallLines,
-        `}`,
-      ].join('\n'),
-    );
-  }
-
-  const sortedTypeNames = [...importedTypeNames].sort();
-  const helperTypeNames = new Set(sortedTypeNames);
-  const combinedValidationErrorTypeName = inlineHelpers.has('validateCombinedRequest')
-    ? reserveUniqueClientTypeIdentifier('PortableValidationError', helperTypeNames)
-    : null;
-  const combinedValidationResultTypeName = inlineHelpers.has('validateCombinedRequest')
-    ? reserveUniqueClientTypeIdentifier('PortableValidationResult', helperTypeNames)
-    : null;
-  const lines = [
-    `import {`,
-    `\tcallEndpoint,`,
-    `\tcreateEndpoint,`,
-    `\ttype EndpointCallOptions,`,
-    ...(inlineHelpers.has('validateCombinedRequest')
-      ? [
-          `\ttype ValidationError as ${combinedValidationErrorTypeName},`,
-          `\ttype ValidationResult as ${combinedValidationResultTypeName},`,
-        ]
-      : []),
-    `} from '@wp-typia/api-client';`,
-    ...(sortedTypeNames.length === 1
-      ? [
-          `import type { ${sortedTypeNames[0]} } from ${toJavaScriptStringLiteral(
-            toModuleImportPath(clientPath, path.resolve(projectRoot, typesFile)),
-          )};`,
-        ]
-      : [
-          `import type {`,
-          ...sortedTypeNames.map((typeName) => `\t${typeName},`),
-          `} from ${toJavaScriptStringLiteral(
-            toModuleImportPath(clientPath, path.resolve(projectRoot, typesFile)),
-          )};`,
-        ]),
-    `import { apiValidators } from ${toJavaScriptStringLiteral(
-      toModuleImportPath(clientPath, path.resolve(projectRoot, validatorsFile)),
-    )};`,
-    '',
-    ...(inlineHelpers.has('validateNoRequest')
-      ? [
-          `function validateNoRequest(input: unknown) {`,
-          `\tif (input !== undefined) {`,
-          `\t\treturn {`,
-          `\t\t\tdata: undefined,`,
-          `\t\t\terrors: [`,
-          `\t\t\t\t{`,
-          `\t\t\t\t\texpected: 'undefined',`,
-          `\t\t\t\t\tpath: '(root)',`,
-          `\t\t\t\t\tvalue: input,`,
-          `\t\t\t\t},`,
-          `\t\t\t],`,
-          `\t\t\tisValid: false,`,
-          `\t\t};`,
-          `\t}`,
-          '',
-          `\treturn {`,
-          `\t\tdata: undefined,`,
-          `\t\terrors: [],`,
-          `\t\tisValid: true,`,
-          `\t};`,
-          `}`,
-          '',
-        ]
-      : []),
-    ...(inlineHelpers.has('validateCombinedRequest')
-      ? [
-          `function validateCombinedRequest<TQuery, TBody>(`,
-          `\tinput: unknown,`,
-          `\tvalidateQuery: (input: unknown) => ${combinedValidationResultTypeName}<TQuery>,`,
-          `\tvalidateBody: (input: unknown) => ${combinedValidationResultTypeName}<TBody>,`,
-          `): ${combinedValidationResultTypeName}<{ query: TQuery; body: TBody }> {`,
-          `\tif ( input === null || typeof input !== 'object' || Array.isArray( input ) ) {`,
-          `\t\treturn {`,
-          `\t\t\tdata: undefined,`,
-          `\t\t\terrors: [`,
-          `\t\t\t\t{`,
-          `\t\t\t\t\texpected: '{ query, body }',`,
-          `\t\t\t\t\tpath: '(root)',`,
-          `\t\t\t\t\tvalue: input,`,
-          `\t\t\t\t},`,
-          `\t\t\t],`,
-          `\t\t\tisValid: false,`,
-          `\t\t};`,
-          `\t}`,
-          ``,
-          `\tconst request = input as { query?: unknown; body?: unknown };`,
-          `\tif ( !Object.prototype.hasOwnProperty.call( request, 'query' ) || !Object.prototype.hasOwnProperty.call( request, 'body' ) ) {`,
-          `\t\treturn {`,
-          `\t\t\tdata: undefined,`,
-          `\t\t\terrors: [`,
-          `\t\t\t\t{`,
-          `\t\t\t\t\texpected: '{ query, body }',`,
-          `\t\t\t\t\tpath: '(root)',`,
-          `\t\t\t\t\tvalue: input,`,
-          `\t\t\t\t},`,
-          `\t\t\t],`,
-          `\t\t\tisValid: false,`,
-          `\t\t};`,
-          `\t}`,
-          ``,
-          `\tconst prefixPath = (prefix: '$.query' | '$.body', path: string): string => {`,
-          `\t\tif ( path === '(root)' ) {`,
-          `\t\t\treturn prefix;`,
-          `\t\t}`,
-          ``,
-          `\t\treturn path.startsWith( '$' ) ? \`\${prefix}\${path.slice( 1 )}\` : \`\${prefix}.\${path}\`;`,
-          `\t};`,
-          ``,
-          `\tconst queryValidation = validateQuery( request.query );`,
-          `\tconst bodyValidation = validateBody( request.body );`,
-          `\tconst errors: ${combinedValidationErrorTypeName}[] = [`,
-          `\t\t...queryValidation.errors.map( ( error ) => ( {`,
-          `\t\t\t...error,`,
-          `\t\t\tpath: prefixPath( '$.query', error.path ),`,
-          `\t\t} ) ),`,
-          `\t\t...bodyValidation.errors.map( ( error ) => ( {`,
-          `\t\t\t...error,`,
-          `\t\t\tpath: prefixPath( '$.body', error.path ),`,
-          `\t\t} ) ),`,
-          `\t];`,
-          ``,
-          `\tif ( !queryValidation.isValid || !bodyValidation.isValid ) {`,
-          `\t\treturn {`,
-          `\t\t\tdata: undefined,`,
-          `\t\t\terrors,`,
-          `\t\t\tisValid: false,`,
-          `\t\t};`,
-          `\t}`,
-          ``,
-          `\treturn {`,
-          `\t\tdata: {`,
-          `\t\t\tbody: bodyValidation.data ?? ( request.body as TBody ),`,
-          `\t\t\tquery: queryValidation.data ?? ( request.query as TQuery ),`,
-          `\t\t},`,
-          `\t\terrors: [],`,
-          `\t\tisValid: true,`,
-          `\t};`,
-          `}`,
-          '',
-        ]
-      : []),
-    ...endpointLines.flatMap((entry) => [entry, '']),
-  ];
-
-  reconcileGeneratedArtifacts(
-    [
-      {
-        content: `${lines.join('\n').trimEnd()}\n`,
-        path: clientPath,
-      },
-    ],
-    executionOptions,
-  );
-
-  return {
-    clientPath,
-    endpointCount: manifest.endpoints.length,
-    operationIds: [...operationIds],
-  };
+  return syncEndpointClientModule(options, executionOptions);
 }
